TextClip 및 crop 정수 변환 오류 수정
- TextClip size 파라미터: video.w * 0.9 → int(video.w * 0.9)
- crop 파라미터: x_center, y_center, width, height 모두 정수로 변환
- PIL/MoviePy는 크기 값으로 정수만 허용

+++ local_cli/services/video_producer.py
@@ -327,7 +327,7 @@ class VideoProducer:
                 stroke_color='black',
                 stroke_width=2,
                 method='caption',
-                size=(video.w * 0.9, None)
+                size=(int(video.w * 0.9), None)
             )
 
         subtitle_clips = []
@@ -342,10 +342,10 @@ class VideoProducer:
         # 숏폼은 9:16 크롭
         if video_format == 'short':
             video = video.cropped(
-                x_center=video.w/2,
-                y_center=video.h/2,
+                x_center=int(video.w/2),
+                y_center=int(video.h/2),
                 width=int(video.h * 9/16),
-                height=video.h
+                height=int(video.h)
             )
 
         return video
